refactor: Move ibis type mappings to compile module
