+++ bigframes/core/compile/aggregate_compiler.py
@@ -22,10 +22,10 @@ import ibis.expr.types as ibis_types
 import pandas as pd
 
 import bigframes.constants as constants
+import bigframes.core.compile.ibis_types as compile_ibis_types
 import bigframes.core.compile.scalar_op_compiler as scalar_compilers
 import bigframes.core.expression as ex
 import bigframes.core.window_spec as window_spec
-import bigframes.dtypes as dtypes
 import bigframes.operations.aggregations as agg_ops
 
 scalar_compiler = scalar_compilers.scalar_op_compiler
@@ -323,7 +323,7 @@ def _(
             for this_bin in range(op.bins - 1):
                 out = out.when(
                     x <= (col_min + (this_bin + 1) * bin_width),
-                    dtypes.literal_to_ibis_scalar(
+                    compile_ibis_types.literal_to_ibis_scalar(
                         this_bin, force_dtype=pd.Int64Dtype()
                     ),
                 )
@@ -352,8 +352,8 @@ def _(
                     out = out.when(x.notnull(), interval_struct)
     else:  # Interpret as intervals
         for interval in op.bins:
-            left = dtypes.literal_to_ibis_scalar(interval[0])
-            right = dtypes.literal_to_ibis_scalar(interval[1])
+            left = compile_ibis_types.literal_to_ibis_scalar(interval[0])
+            right = compile_ibis_types.literal_to_ibis_scalar(interval[1])
             condition = (x > left) & (x <= right)
             interval_struct = ibis.struct(
                 {"left_exclusive": left, "right_inclusive": right}
@@ -370,7 +370,7 @@ def _(
     window=None,
 ) -> ibis_types.IntegerValue:
     if isinstance(self.quantiles, int):
-        quantiles_ibis = dtypes.literal_to_ibis_scalar(self.quantiles)
+        quantiles_ibis = compile_ibis_types.literal_to_ibis_scalar(self.quantiles)
         percent_ranks = cast(
             ibis_types.FloatingColumn,
             _apply_window_if_present(column.percent_rank(), window),
@@ -383,13 +383,19 @@ def _(
             _apply_window_if_present(column.percent_rank(), window),
         )
         out = ibis.case()
-        first_ibis_quantile = dtypes.literal_to_ibis_scalar(self.quantiles[0])
+        first_ibis_quantile = compile_ibis_types.literal_to_ibis_scalar(
+            self.quantiles[0]
+        )
         out = out.when(percent_ranks < first_ibis_quantile, None)
         for bucket_n in range(len(self.quantiles) - 1):
-            ibis_quantile = dtypes.literal_to_ibis_scalar(self.quantiles[bucket_n + 1])
+            ibis_quantile = compile_ibis_types.literal_to_ibis_scalar(
+                self.quantiles[bucket_n + 1]
+            )
             out = out.when(
                 percent_ranks <= ibis_quantile,
-                dtypes.literal_to_ibis_scalar(bucket_n, force_dtype=pd.Int64Dtype()),
+                compile_ibis_types.literal_to_ibis_scalar(
+                    bucket_n, force_dtype=pd.Int64Dtype()
+                ),
             )
         out = out.else_(None)
         return out.end()  # type: ignore
