+++ bigframes/core/compile/compiled.py
@@ -28,6 +28,7 @@ import ibis.expr.types as ibis_types
 import pandas
 
 import bigframes.core.compile.aggregate_compiler as agg_compiler
+import bigframes.core.compile.ibis_types
 import bigframes.core.compile.scalar_op_compiler as op_compilers
 import bigframes.core.expression as ex
 import bigframes.core.guid
@@ -157,16 +158,19 @@ class BaseIbisIR(abc.ABC):
             )
         return typing.cast(
             ibis_types.Value,
-            bigframes.dtypes.ibis_value_to_canonical_type(self._column_names[key]),
+            bigframes.core.compile.ibis_types.ibis_value_to_canonical_type(
+                self._column_names[key]
+            ),
         )
 
     def get_column_type(self, key: str) -> bigframes.dtypes.Dtype:
         ibis_type = typing.cast(
-            bigframes.dtypes.IbisDtype, self._get_ibis_column(key).type()
+            bigframes.core.compile.ibis_types.IbisDtype,
+            self._get_ibis_column(key).type(),
         )
         return typing.cast(
             bigframes.dtypes.Dtype,
-            bigframes.dtypes.ibis_dtype_to_bigframes_dtype(ibis_type),
+            bigframes.core.compile.ibis_types.ibis_dtype_to_bigframes_dtype(ibis_type),
         )
 
     def _aggregate_base(
@@ -332,7 +336,8 @@ class UnorderedIR(BaseIbisIR):
         # Make sure all dtypes are the "canonical" ones for BigFrames. This is
         # important for operations like UNION where the schema must match.
         table = self._table.select(
-            bigframes.dtypes.ibis_value_to_canonical_type(column) for column in columns
+            bigframes.core.compile.ibis_types.ibis_value_to_canonical_type(column)
+            for column in columns
         )
         base_table = table
         if self._reduced_predicate is not None:
@@ -579,7 +584,10 @@ class OrderedIR(BaseIbisIR):
         ibis_values = ibis_values.assign(**{ORDER_ID_COLUMN: range(len(pd_df))})
         # derive the ibis schema from the original pandas schema
         ibis_schema = [
-            (name, bigframes.dtypes.bigframes_dtype_to_ibis_dtype(dtype))
+            (
+                name,
+                bigframes.core.compile.ibis_types.bigframes_dtype_to_ibis_dtype(dtype),
+            )
             for name, dtype in zip(schema.names, schema.dtypes)
         ]
         ibis_schema.append((ORDER_ID_COLUMN, ibis_dtypes.int64))
@@ -993,7 +1001,9 @@ class OrderedIR(BaseIbisIR):
         # Make sure all dtypes are the "canonical" ones for BigFrames. This is
         # important for operations like UNION where the schema must match.
         table = table.select(
-            bigframes.dtypes.ibis_value_to_canonical_type(table[column])
+            bigframes.core.compile.ibis_types.ibis_value_to_canonical_type(
+                table[column]
+            )
             for column in table.columns
         )
         base_table = table
